Extension v2
What happens right now is that the correct url is passed but no action is taken on the page (idk why).

I guess contentscript doesn't do anything except of sending the url, and the rules doesn't get updated properly, or maybe it is Ublock.

Next thing I need to do is to retrain the model, I added a lot of benign domains so hopefully it doesn't recognize google.com as malware.