Enhancements
Added the route to retrain the model, not implemented in the option page.

Warning: retraining takes the whole dataset and add the whitelisted elements at the end, so the training takes a lot since it is on CPU, it is better to retrain the model at zero since training it with the pre-trained weights induce unwanted behaviour and no actual benefit.

Added stuff inside the Option page: see logs and whitelisted elements.

To-do: Add the retrain button with a lot of warnings in the option page.